Fix vlv CI test deadlock
@long-entryrdn was not open by dbmdb_open_all_files
this leaded to failure when trying to open it in a read operation
because at dblayer level, it is not possible to open write txn
within a read txn - and although it is possible at lmdb level,
the new file will not be visible within the read txn
but we may need to access it.
So the open failed, and entryrdn attrinfo should then be released
before returning an error to avoid keeping entryrdn busy.
That is what trigger the hang when removing a backend.

Added some conditionnal debug code to understand why the server hang.

Also added a missing dblayer_release_index_file in vlvIndex_checkforindex
that may be the reason while there is a hang when removing vlv on bdb.

+++ dirsrvtests/tests/suites/vlv/regression_test.py
@@ -186,7 +186,7 @@ def vlv_setup_with_two_backend(topology_st, request):
 
     def fin():
         # Cleanup function
-        if not DEBUGGING:
+        if not DEBUGGING and inst.exists() and inst.status():
             for be in Backends(inst).list():
                 if be.get_attr_val_utf8_l('cn') in [ 'be1', 'be2' ]:
                     be.delete()
@@ -826,6 +826,7 @@ def test_vlv_as_freeipa_backup_restore(freeipa):
     DATE_LIMIT = '20260323135138Z'
     EXPECTED_COUNT = 12
     inst = freeipa.standalone
+    inst.restart()
     conn = open_new_ldapi_conn(inst.serverid)
     basedn = 'ou=certificateRepository,ou=ca,o=ipaca'
     count = len(conn.search_s(basedn, ldap.SCOPE_SUBTREE, f"(&(certStatus=VALID)(notAfter<={DATE_LIMIT}))"))
